+++ src/form/b-icon-button/b-icon-button.ts
@@ -6,9 +6,17 @@
  * https://github.com/V4Fire/Client/blob/master/LICENSE
  */
 
+/**
+ * [[include:form/b-icon-button/README.md]]
+ * @packageDocumentation
+ */
+
 import bButton, { component } from 'form/b-button/b-button';
 
 export * from 'form/b-button/b-button';
 
+/**
+ * Component to create a button based on icon
+ */
 @component({flyweight: true})
 export default class bIconButton extends bButton {}