+++ src/form/b-button/interface.ts
@@ -10,4 +10,5 @@ export type ButtonType<T extends string = any> =
 	'submit' |
 	'button' |
 	'link' |
+	'file' |
 	T;