+++ CHANGELOG.md
@@ -16,6 +16,7 @@ _Note: Gaps between patch versions are faulty, broken or test releases._
 #### :boom: Breaking Change
 
 * The component was renamed `b-input-birthday` -> `b-select-date`
+* Removed `bFileButton`
 
 #### :rocket: New Feature